prompt_tracking_package/requirements.txt
pandas
python-dotenv
google-generativeai
transformers
stanza
unittest
